requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.2.5
django-environ==0.11.2
django-js-asset==2.1.0
django-mptt==0.14.0
Pillow==10.0.1
sqlparse==0.4.4
psycopg2-binary==2.9.7
gunicorn==21.2.0
django-honeypot==1.0.3
django-phonenumber-field==7.1.0
phonenumbers==8.13.21